// src/app/api/model/planning-solution-view-model.ts
/**
 * Conference Schedule Optimizer API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TimeslotViewModel } from './timeslot-view-model';
import { TalkViewModel } from './talk-view-model';
import { ConstraintMatchViewModel } from './constraint-match-view-model';
import { ScoreViewModel } from './score-view-model';


export interface PlanningSolutionViewModel { 
    id: number;
    score?: ScoreViewModel;
    constraintMatches?: Array<ConstraintMatchViewModel>;
    talks: Array<TalkViewModel>;
    timeslots: Array<TimeslotViewModel>;
}