// src/app/api/model/timeslot-view-model.ts
/**
 * Conference Schedule Optimizer API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TimeslotViewModel { 
    id: number;
    room: string;
    /**
     * Start time of the timeslot (HH:mm)
     */
    start?: string;
    /**
     * End time of the timeslot (HH:mm)
     */
    end?: string;
    seats: number;
    beamerAvailable?: boolean;
    flipchartAvailable?: boolean;
}